pyproject.toml
[tool.poetry]
name = "grafikalaby"
version = "0.1.0"
description = ""
authors = ["yano358 <jan.gumulakdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyOpenGL = "^3.1.7"
PyOpenGL_accelerate = "^3.1.7"
setuptools = "^65.6.3"
wheel = "^0.38.4"
glfw = "^2.6.2"
python-cowsay = "^1.1.1" 
numpy = "^1.26.2"
pyglm = "^2.7.1"
pillow = "^10.1.0"
matplotlib = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
